Tests/LibWeb/Text/expected/WebAnimations/animation-methods/finished-fill-forwards-animations-on-different-elements-are-not-replaced.txt

first animations after timeout: 1
first opacity after timeout: 0
second animations after timeout: 1
second opacity after timeout: 0
